+++ Src/PPEquip/EquipDll/CustomDispaly/Viki_Vision.cpp
@@ -68,7 +68,7 @@ int VikiVsnEquip::GetConfig(char * pBuf, size_t bufSize)
 {
 	int ok = 0;
 	SString str;
-	str.CatEq("STRLEN", (long)STRLEN);
+	str.CatEq("STRLEN", STRLEN);
 	if(str.Len() + 1 > bufSize)
 		ok = 2;
 	else
