+++ Src/PPLib/ppserver.cpp
@@ -4494,7 +4494,7 @@ int run_client()
 						else if(r == 1) {
 							const PPJobSrvProtocol::Header & hdr = reply.GetH();
 							reply_str.Z().Cat("Binary reply").CatDiv(':', 2).
-								CatEq("ProtocolVer", (long)hdr.ProtocolVer).CatDiv(';', 2).
+								CatEq("ProtocolVer", hdr.ProtocolVer).CatDiv(';', 2).
 								CatEq("Padding", (uint)hdr.Padding).CatDiv(';', 2). // @v11.0.10
 								CatEq("DataLen", hdr.DataLen).CatDiv(';', 2).
 								CatEq("DataType", hdr.Type).CatDiv(';', 2).
